feat: Updates login/signup to Relay flow
This updates the login/signup process to a Relay based flow.
Breaking changes here, as you now must duplicate the `JWT_PRIVATE_KEY`
environment variable from the api repo. Both the API and frontend now
require this. The front end requires it so that it can sign an anonymous
JWT token that is used to make queries to the database under the
`postgraphql_anonymous` role.