feat: Update login and signup for Auth
This fixes the login and signup pages to use the new payload that is
being returned from the GraphQL server. It is now returning a user_id
and a signed jwt token on login and signup that has been JSON
stringified. This changes the code and tests to handle that.